Fix Burdock::Actions::ScopeAtLine::Processor handling of begin and module
The `#on_module` handling was incorrect as it was not updating the
`:module` node's child nodes with the result of calling `#process` on
the node uncovered by `#find_node_on_line`.

The `#on_begin` handling was not necessarily incorrect but was altered
to return only the node uncovered by `#find_node_on_line` instead of
updating the `:begin` node itself. This decision was motivated by
testing as it became obvious that accounting for the AST `begin` node, a
concept entirely of AST, might require more effort.